+++ CHANGELOG.rst
@@ -21,6 +21,9 @@ Features:
 * Added a ``marker`` option to the ``plot_iso`` function
   which acts similar to the ``color`` parameter and allows
   simple selection of the marker style.
+* Added two new isotherm models: Dubinin-Radushkevich and
+  Dubinin-Astakov. They can be used for fitting by specifying
+  `DR` and `DA` as the model, respectivelly.
 
 Breaking changes: